requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.44.0
django-cors-headers==3.8.0
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
msgpack==1.0.2
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zope.interface==5.4.0
